requirements.txt
flask
requests
python-binance
psycopg2-binary
websocket-client
pandas
tqdm